dalhe

+++ src/Models/Entities/Movie.cs
@@ -13,7 +13,7 @@ namespace API.Models.Entities
 
         public string Synopsis { get; set; }
 
-        public float Duration { get; set; }  
+        public int Duration { get; set; }  
 
         public MovieGenre Genre { get; set; }
