+++ src/Models/UpdatedEntity/UpdatedMovie.cs
@@ -14,7 +14,7 @@ namespace API.Models.UpdatedEntity
 
         public string? Synopsis { get; set; }
 
-        public float? Duration { get; set; }
+        public int? Duration { get; set; }
 
         public MovieGenre? Genre { get; set; }
